rtx5090/scoping_review/requirements.txt
# Scoping review automation deps (on top of ../requirements.txt)
biopython>=1.81
requests>=2.31
rapidfuzz>=3.5
tldextract>=5.1
tqdm>=4.66
pandas>=2.0
pyyaml>=6.0
anthropic>=0.34
ollama>=0.2  # optional local LLM client
